+++ packages/patternfly-3/react-console/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.5.2"></a>
+## [1.5.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-console@1.5.1...@patternfly/react-console@1.5.2) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package @patternfly/react-console
+
 <a name="1.5.1"></a>
 ## [1.5.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-console@1.5.0...@patternfly/react-console@1.5.1) (2018-10-29)