+++ packages/patternfly-4/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.15.3"></a>
+## [1.15.3](https://github.com/patternfly/patternfly-react/compare/patternfly4-react-lerna-root@1.15.2...patternfly4-react-lerna-root@1.15.3) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package patternfly4-react-lerna-root
+
+<a name="1.15.2"></a>
+## [1.15.2](https://github.com/patternfly/patternfly-react/compare/patternfly4-react-lerna-root@1.15.1...patternfly4-react-lerna-root@1.15.2) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package patternfly4-react-lerna-root
+
 <a name="1.15.1"></a>
 ## [1.15.1](https://github.com/patternfly/patternfly-react/compare/patternfly4-react-lerna-root@1.15.0...patternfly4-react-lerna-root@1.15.1) (2018-10-29)