+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.28.3"></a>
+## [1.28.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.28.2...@patternfly/react-core@1.28.3) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package @patternfly/react-core
+
+<a name="1.28.2"></a>
+## [1.28.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.28.1...@patternfly/react-core@1.28.2) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package @patternfly/react-core
+
 <a name="1.28.1"></a>
 ## [1.28.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.28.0...@patternfly/react-core@1.28.1) (2018-10-29)